+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## [1.12] -
 - support GHC 8.8, 8.10, 9.0 only
+- add patterns for each month of year
 - fix: don't provide TAI clock where it's unavailable (e.g. FreeBSD)
 - fix: handle time of day 24:00:00 for ISO 8601 parsing (only)
 - fix parsing of %f and %G with negative years